+++ apps/chat/src/components/Header/Header.tsx
@@ -1,5 +1,5 @@
 import { memo } from 'react';
-import UserMenu from '../auth/UserMenu';
+import UserMenu from './UserMenu';
 import Logo from './Logo';
 
 const Header = () => {
@@ -8,6 +8,7 @@ const Header = () => {
       <div className="flex flex-1 justify-center">
         <Logo />
       </div>
+      {/* TODO: remove - need to be in navigation panel */}
       <div className="absolute right-2">
         <UserMenu />
       </div>
